+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Types.hs
@@ -79,6 +79,7 @@ import Cardano.Ledger.Shelley.LedgerState as X (
   CertState (..),
   DState (..),
   EpochState (..),
+  VState (..),
   IncrementalStake (..),
   InstantaneousRewards (..),
   KeyPairs, -- deprecated
@@ -88,7 +89,6 @@ import Cardano.Ledger.Shelley.LedgerState as X (
   RewardUpdate (..),
   ShelleyPPUPState (..),
   UTxOState (..),
-  VState (..),
  )
 import Cardano.Ledger.Shelley.PParams as X (
   ProposedPPUpdates (..),