+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/ModelState.hs
@@ -70,6 +70,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   CertState (..),
   DState (..),
   EpochState (..),
+  VState (..),
   IncrementalStake (..),
   InstantaneousRewards (..),
   LedgerState (..),
@@ -77,7 +78,6 @@ import Cardano.Ledger.Shelley.LedgerState (
   PState (..),
   StashedAVVMAddresses,
   UTxOState (..),
-  VState (..),
   completeRupd,
   smartUTxOState,
  )